[FIX] 2.5차 분리
