+++ operation-api/src/main/java/org/operation/common/config/TimezoneConfig.java
@@ -1,4 +1,4 @@
-package org.sopt.makers.operation.config;
+package org.operation.common.config;
 
 import java.util.TimeZone;
 
@@ -7,7 +7,7 @@ import javax.annotation.PostConstruct;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-public class TimezoneConfig {
+public class TimezoneConfig { //TODO: 타임 체크 로컬에서 UTF로 바꿔서 체크
 
 	@PostConstruct
 	public void init() {
