+++ operation-business/build.gradle
@@ -1,0 +1,22 @@
+plugins {
+    id 'java'
+}
+
+group = 'org.sopt.makers'
+version = '0.0.1-SNAPSHOT'
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    implementation project(path: ':operation-api')
+    implementation project(path: ':operation-domain')
+
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+}
+
+test {
+    useJUnitPlatform()
+}
