+++ operation-business/src/main/java/org/operation/common/exception/AdminFailureException.java
@@ -1,4 +1,4 @@
-package org.sopt.makers.operation.exception;
+package org.operation.common.exception;
 
 public class AdminFailureException extends RuntimeException {
     public AdminFailureException(String message) {
