+++ operation-business/src/main/java/org/operation/common/exception/AlarmException.java
@@ -1,4 +1,4 @@
-package org.sopt.makers.operation.exception;
+package org.operation.common.exception;
 
 public class AlarmException extends RuntimeException {
     public AlarmException(String message) {
