+++ operation-business/src/main/java/org/operation/common/exception/LectureException.java
@@ -1,4 +1,4 @@
-package org.sopt.makers.operation.exception;
+package org.operation.common.exception;
 
 public class LectureException extends RuntimeException {
     public LectureException(String message) {
