+++ operation-business/src/main/java/org/operation/common/exception/TokenException.java
@@ -1,4 +1,4 @@
-package org.sopt.makers.operation.exception;
+package org.operation.common.exception;
 public class TokenException extends RuntimeException {
     public TokenException(String message) {
         super("[TokenException] : " + message);
